Move refactor, cleanup unused util
